Re-wrote to avoid platforms in settings
