+++ src/main/ipc-handlers.ts
@@ -1,12 +1,11 @@
 import { app, dialog, ipcMain, shell } from 'electron';
 import { platform } from 'os';
 import { enumerateValues, enumerateKeys } from 'registry-js';
-import { GamePlatform, PlatformRunType } from '../common/GamePlatform';
+import { DefaultGamePlatforms, GamePlatform, PlatformRunType } from '../common/GamePlatform';
 import spawn from 'cross-spawn';
 import path from 'path';
 
 import { IpcMessages, IpcOverlayMessages } from '../common/ipc-messages';
-import { GamePlatformInstance, GamePlatformMap } from '../common/ISettings';
 
 // Listeners are fire and forget, they do not have "responses" or return values
 export const initializeIpcListeners = (): void => {
@@ -16,50 +15,9 @@ export const initializeIpcListeners = (): void => {
 		}
 	});
 
-	ipcMain.handle(IpcMessages.REQUEST_PLATFORMS_AVAILABLE, (_, platforms: GamePlatformMap) => {
-		
-		const desktop_platform = platform();
+	ipcMain.on(IpcMessages.OPEN_AMONG_US_GAME, (_, platformKey: GamePlatform) => {
 
-		// Assume all platforms are false unless proven otherwise
-		for (const key in platforms) {
-			const game_platform = platforms[key];
-			game_platform.available = false;
-
-			if (desktop_platform === 'win32') {
-			
-				if (game_platform.key === GamePlatform.EPIC || game_platform.key === GamePlatform.STEAM) {
-					// Search registry for the URL Protocol
-					if (enumerateValues(game_platform.registryKey, game_platform.registrySubKey).find(
-						(value) => value ? value.name === game_platform.registryKeyValue : false
-					)) {
-						game_platform.available = true;
-					}
-				} else if (game_platform.key === GamePlatform.MICROSOFT) {
-					// Search for 'Innersloth.Among Us....' key and grab it
-					const key_found = enumerateKeys(game_platform.registryKey, game_platform.registrySubKey).find(
-						(reg_key) => reg_key.startsWith(game_platform.registryFindKey as string));
-					
-					if (key_found) {
-						// Grab the game path from the above key
-						const value_found = enumerateValues(game_platform.registryKey, game_platform.registrySubKey + '\\' + key_found).find(
-							(value) => value ? value.name === game_platform.registryKeyValue : false
-						);
-						if (value_found) {
-							game_platform.available = true;
-							game_platform.run = value_found.data as string;
-						}
-					}
-				}
-			} else if (desktop_platform === 'linux') {
-				// TODO: Platform checking on Linux
-				// Set 'game_platform.available' true and setup data if platform is available, do nothing otherwise
-				continue;
-			}
-		}
-		return platforms;
-	});
-
-	ipcMain.on(IpcMessages.OPEN_AMONG_US_GAME, (_, platform: GamePlatformInstance) => {
+		const platform = DefaultGamePlatforms[platformKey];
 
 		const error = () => dialog.showErrorBox('Error', 'Could not start the game.');
 		
@@ -120,5 +78,46 @@ export const initializeIpcListeners = (): void => {
 // or the caller should be "await"'ing them.  If neither of these are the case
 // consider making it a "listener" instead for performance and readability
 export const initializeIpcHandlers = (): void => {
-	// TODO: Put handlers here
+	ipcMain.handle(IpcMessages.REQUEST_PLATFORMS_AVAILABLE, () => {
+		
+		const desktop_platform = platform();
+
+		// Assume all platforms are false unless proven otherwise
+		for (const key in DefaultGamePlatforms) {
+			const game_platform = DefaultGamePlatforms[key];
+			game_platform.available = false;
+
+			if (desktop_platform === 'win32') {
+			
+				if (game_platform.key === GamePlatform.EPIC || game_platform.key === GamePlatform.STEAM) {
+					// Search registry for the URL Protocol
+					if (enumerateValues(game_platform.registryKey, game_platform.registrySubKey).find(
+						(value) => value ? value.name === game_platform.registryKeyValue : false
+					)) {
+						game_platform.available = true;
+					}
+				} else if (game_platform.key === GamePlatform.MICROSOFT) {
+					// Search for 'Innersloth.Among Us....' key and grab it
+					const key_found = enumerateKeys(game_platform.registryKey, game_platform.registrySubKey).find(
+						(reg_key) => reg_key.startsWith(game_platform.registryFindKey as string));
+					
+					if (key_found) {
+						// Grab the game path from the above key
+						const value_found = enumerateValues(game_platform.registryKey, game_platform.registrySubKey + '\\' + key_found).find(
+							(value) => value ? value.name === game_platform.registryKeyValue : false
+						);
+						if (value_found) {
+							game_platform.available = true;
+							game_platform.run = value_found.data as string;
+						}
+					}
+				}
+			} else if (desktop_platform === 'linux') {
+				// TODO: Platform checking on Linux
+				// Set 'game_platform.available' true and setup data if platform is available, do nothing otherwise
+				continue;
+			}
+		}
+		return DefaultGamePlatforms;
+	});
 };
