+++ src/renderer/App.tsx
@@ -35,7 +35,7 @@ import 'typeface-varela/index.css';
 import { DEFAULT_PLAYERCOLORS } from '../main/avatarGenerator';
 import './language/i18n';
 import { withNamespaces } from 'react-i18next';
-import { DefaultGamePlatforms, GamePlatform } from '../common/GamePlatform';
+import { GamePlatform } from '../common/GamePlatform';
 let appVersion = '';
 if (typeof window !== 'undefined' && window.location) {
 	const query = new URLSearchParams(window.location.search.substring(1));
@@ -174,7 +174,6 @@ export default function App({ t }): JSX.Element {
 			publicLobby_language: 'en',
 			publicLobby_mods: 'NONE',
 		},
-		launchPlatformSettings: DefaultGamePlatforms,
 		launchPlatform: GamePlatform.STEAM,
 	});
 	const lobbySettings = useReducer(lobbySettingsReducer, settings[0].localLobbySettings);
